# apps/landing/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fluxby API",
    "version": "1.0.0",
    "description": "API for managing financial transactions, categories, budgets, and analytics. This API is intended for developers building custom interfaces.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/fluxby/fluxby/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/api/accounts": {
      "get": {
        "summary": "Haal alle rekeningen op",
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Profile-Id",
            "schema": {
              "type": "integer"
            },
            "description": "Profile ID to filter accounts by"
          }
        ],
        "responses": {
          "200": {
            "description": "Lijst met rekeningen inclusief saldo"
          }
        }
      },
      "post": {
        "summary": "Maak een nieuwe rekening aan",
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Profile-Id",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "iban",
                  "name"
                ],
                "properties": {
                  "iban": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "checking",
                      "savings",
                      "credit"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bestaande rekening bijgewerkt"
          },
          "201": {
            "description": "Rekening aangemaakt"
          },
          "400": {
            "description": "IBAN en naam zijn verplicht"
          }
        }
      },
      "delete": {
        "summary": "Delete all accounts for current profile",
        "description": "Deletes all accounts for the current profile. Transactions will be unlinked (account_id set to null) but not deleted.",
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Profile-Id",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Accounts deleted"
          }
        }
      }
    },
    "/api/accounts/{id}": {
      "get": {
        "summary": "Haal een specifieke rekening op",
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Profile-Id",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rekening details"
          },
          "404": {
            "description": "Rekening niet gevonden"
          }
        }
      },
      "patch": {
        "summary": "Werk een rekening bij",
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Profile-Id",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "checking",
                      "savings",
                      "credit"
                    ]
                  },
                  "currentBalance": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rekening succesvol bijgewerkt"
          },
          "403": {
            "description": "Access denied"
          }
        }
      },
      "delete": {
        "summary": "Delete an account",
        "description": "Deletes an account. Transactions linked to this account will remain but be unlinked (account_id set to null).",
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Profile-Id",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account deleted"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/api/accounts/order": {
      "patch": {
        "summary": "Update the display order of accounts",
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Profile-Id",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "accountIds"
                ],
                "properties": {
                  "accountIds": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "description": "Array of account IDs in the desired display order"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account order updated successfully"
          },
          "400": {
            "description": "Invalid request data"
          }
        }
      }
    },
...
  },
...
}